Move state to App component for state persistence
